requirement.txt
SpeechRecognition==3.10.0
PyAudio==0.2.11
pyttsx3==2.90
datetime
webbrowser
os
subprocess
pyautogui==0.9.53
requests==2.31.0
psutil==5.9.6
ctypes
random
wikipedia==1.4.0
json
platform
shutil
time
socket
speedtest-cli==2.1.3
spotipy==2.23.0
musilib  